even more binary trees!
I wish I understood why the second solution for max depth is faster than the first. They do exactly the same thing, one is just condensed to a single line thanks to expression body and ternary. They both check if the input is null then return 0 if true or 1+ max of recursive call for each node